+++ test/integration/migrations-test.js
@@ -1,9 +1,9 @@
 const knexMigrate = require('knex-migrate');
 
 async function migrateToLatest(knexFilePath) {
-  const log = ({ action, migration }) => console.log('Doing ' + action + ' on ' + migration);
+  const log = ({ action, migration }) => console.log(`Doing ${  action  } on ${  migration}`);
 
-  console.log('knexfile: ' + knexFilePath);
+  console.log(`knexfile: ${  knexFilePath}`);
 
   await knexMigrate(
     'up',
@@ -11,10 +11,10 @@ async function migrateToLatest(knexFilePath) {
       knexfile: knexFilePath,
       migrations: 'test/integration/migrations',
     },
-    log,
+    log
   );
 }
 
 module.exports = {
-  migrateToLatest,
+  migrateToLatest
 };
